notary: Enable CouchDB backend for notary signer and server
Enable the CouchDB backend for the notary signer and server.

The code is mostly derived from the Rethink implementation with the
main difference that we do not require client certificates. This is
because the cloud version of CouchDB, Cloudant, does not require
(honor) client certificates.

Several test cases are added for the CouchDB support.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>