couchdb: server: get rid of sort and limit query over gun+role+version
To improve the query performance of the GetCurrent() operation do not
include an unrestricted version number in the query but restrict the query
to the fields for which we have concrete values, which are 'gun' and 'role'.
Also get rid of the sort and limit in the query but have all the data
returned to the caller and have the caller find the entry with the latest
version.

Restrict the number of fields returned in the response to those that are
needed so that less data needs to be transferred.

This change has had a positive effect on the performance of this particular
operation.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>